requirements.txt
# SonIA Core â Dependencies
# BloomsPal

# Web Framework (for health endpoint + cron trigger)
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AWS DynamoDB (READ ONLY)
boto3==1.34.0

# HTTP Client (for FedEx API, Odoo, SonIA Agent)
httpx==0.26.0

# Scheduler
apscheduler==3.10.4

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Environment variables
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7
